Remove option for caching
