+++ src/main/java/me/phqsh/ns4j/request/http/HttpRequest.java
@@ -1,11 +1,11 @@
-package me.phqsh.ns4j.request;
+package me.phqsh.ns4j.request.http;
 
 import me.phqsh.ns4j.containers.Container;
 
 import java.util.List;
 import java.util.Map;
 
-public interface Request {
+public interface HttpRequest {
     Map<String, List<String>> getResponseHeaders() throws IllegalAccessException;
     Container execute();
     String getUrl();
